+++ src/exodus/hdmint/sigmawallet.cpp
@@ -2,7 +2,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "wallet.h"
+#include "sigmawallet.h"
 
 #include "../exodus.h"
